Midlertidig renaming fra BrevmottakerContext tilbake til VergeContext for å redusere opplevd diff med main
